+++ examples/read_block.ts
@@ -1,6 +1,6 @@
 import * as C from "../mod.ts";
 import * as U from "../util/mod.ts";
 
-const root = C.blockRead(C.polkadot)();
+const root = C.chain.getBlock(C.polkadot)();
 
 console.log(U.throwIfError(await root.run()));
